+++ src/app/common/cv-skill-short.directive.js
@@ -3,10 +3,10 @@
 
     angular
         .module('app.common')
-        .directive('skillShort', skillShortDirective);
+        .directive('cvSkillShort', cvSkillShortDirective);
 
-    function skillShortDirective() {
-        var template = '{{skill.name}} <skill-badge-level level="skill.level"></skill-badge-level> ';
+    function cvSkillShortDirective() {
+        var template = '{{skill.name}} <cv-skill-badge-level level="skill.level"></cv-skill-badge-level> ';
         template += '<a ng-if="(skill.skillUrl != undefined)" href="{{skill.skillUrl}}" title="{{skill.skillUrlTitle}}" target="_blank">';
         template += '   <i class="glyphicon glyphicon-link"></i>';
         template += '</a>';