+++ src/app/modules/home/cv-experience-short.directive.js
@@ -2,10 +2,10 @@
     'use strict';
 
     angular
-        .module('app.workExperiences')
-        .directive('experienceFull', experienceFullDirective);
+        .module('app.home')
+        .directive('cvExperienceShort', cvExperienceShortDirective);
 
-    function experienceFullDirective() {
+    function cvExperienceShortDirective() {
         var template = '<div class="panel panel-default">';
         template += '       <div class="panel-body">';
         template += '           <a href="{{job.companyUrl}}" target="_blank"><img src="{{job.logoUrl}}" alt="{{job.companyUrlTitle}}" class="img-rounded pull-right"></a>';
@@ -13,9 +13,6 @@
         template += '           <h4>{{job.company}}</h4>';
         template += '           <h4><small>{{job.period}}</small></h4>';
         template += '           <p>{{job.jobDescription}}</p>';
-        template += '           <ul>';
-        template += '               <li ng-repeat="task in job.tasks">{{task.desc}}</li>';
-        template += '           </ul>';
         template += '       </div>';
         template += '   </div>';