+++ src/app/modules/work-experiences/experience-full.directive.js
@@ -2,10 +2,10 @@
     'use strict';
 
     angular
-        .module('app.home')
-        .directive('experienceShort', experienceShortDirective);
+        .module('app.workExperiences')
+        .directive('cvExperienceFull', cvExperienceFullDirective);
 
-    function experienceShortDirective() {
+    function cvExperienceFullDirective() {
         var template = '<div class="panel panel-default">';
         template += '       <div class="panel-body">';
         template += '           <a href="{{job.companyUrl}}" target="_blank"><img src="{{job.logoUrl}}" alt="{{job.companyUrlTitle}}" class="img-rounded pull-right"></a>';
@@ -13,6 +13,9 @@
         template += '           <h4>{{job.company}}</h4>';
         template += '           <h4><small>{{job.period}}</small></h4>';
         template += '           <p>{{job.jobDescription}}</p>';
+        template += '           <ul>';
+        template += '               <li ng-repeat="task in job.tasks">{{task.desc}}</li>';
+        template += '           </ul>';
         template += '       </div>';
         template += '   </div>';